Move AssessmentsForPopulationTable and RecordPickerDialog to assessments module
